+++ packages/core/test/tools/sentryBabelTransformer.test.ts
@@ -1,10 +1,13 @@
 import * as process from 'process';
 
-import { SENTRY_DEFAULT_BABEL_TRANSFORMER_PATH } from '../../src/js/tools/sentryBabelTransformerUtils';
+import {
+  createSentryBabelTransformer,
+  SENTRY_BABEL_TRANSFORMER_OPTIONS,
+  SENTRY_DEFAULT_BABEL_TRANSFORMER_PATH,
+} from '../../src/js/tools/sentryBabelTransformerUtils';
 
 process.env[SENTRY_DEFAULT_BABEL_TRANSFORMER_PATH] = require.resolve('./fixtures/mockBabelTransformer.js');
 
-import * as SentryBabelTransformer from '../../src/js/tools/sentryBabelTransformer';
 import type { BabelTransformerArgs } from '../../src/js/tools/vendor/metro/metroBabelTransformer';
 
 const MockDefaultBabelTransformer: {
@@ -15,17 +18,20 @@ const MockDefaultBabelTransformer: {
 describe('SentryBabelTransformer', () => {
   beforeEach(() => {
     jest.clearAllMocks();
+
+    // eslint-disable-next-line @typescript-eslint/no-dynamic-delete
+    delete process.env[SENTRY_BABEL_TRANSFORMER_OPTIONS];
   });
 
   test('getCacheKey calls the original transformer', () => {
-    SentryBabelTransformer.getCacheKey?.();
+    createSentryBabelTransformer().getCacheKey?.();
 
-    expect(SentryBabelTransformer.getCacheKey).toBeDefined();
+    expect(createSentryBabelTransformer().getCacheKey).toBeDefined();
     expect(MockDefaultBabelTransformer.getCacheKey).toHaveBeenCalledTimes(1);
   });
 
   test('transform calls the original transformer with the annotation plugin', () => {
-    SentryBabelTransformer.transform?.({
+    createSentryBabelTransformer().transform?.({
       filename: '/project/file',
       options: {
         projectRoot: 'project/root',
@@ -47,13 +53,51 @@ describe('SentryBabelTransformer', () => {
   });
 
   test('transform adds plugin', () => {
-    SentryBabelTransformer.transform?.({
-      filename: '/project/file',
-      options: {
-        projectRoot: 'project/root',
+    createSentryBabelTransformer().transform?.(createMinimalMockedTransformOptions());
+
+    expect(MockDefaultBabelTransformer.transform).toHaveBeenCalledTimes(1);
+    expect(MockDefaultBabelTransformer.transform).toHaveBeenCalledWith(
+      expect.objectContaining({
+        plugins: expect.arrayContaining([expect.objectContaining({ name: 'componentNameAnnotatePlugin' })]),
+      }),
+    );
+  });
+
+  test('transform adds plugin with options', () => {
+    process.env[SENTRY_BABEL_TRANSFORMER_OPTIONS] = JSON.stringify({
+      annotateReactComponents: {
+        ignoredComponents: ['MyCustomComponent'],
       },
-      plugins: [],
-    } as BabelTransformerArgs);
+    });
+
+    createSentryBabelTransformer().transform?.(createMinimalMockedTransformOptions());
+
+    expect(MockDefaultBabelTransformer.transform).toHaveBeenCalledTimes(1);
+    expect(MockDefaultBabelTransformer.transform).toHaveBeenCalledWith(
+      expect.objectContaining({
+        plugins: expect.arrayContaining([
+          [
+            expect.objectContaining({ name: 'componentNameAnnotatePlugin' }),
+            expect.objectContaining({
+              ignoredComponents: ['MyCustomComponent'],
+            }),
+          ],
+        ]),
+      }),
+    );
+  });
+
+  test('degrades gracefully if options can not be parsed, transform adds plugin without options', () => {
+    process.env[SENTRY_BABEL_TRANSFORMER_OPTIONS] = 'invalid json';
+
+    createSentryBabelTransformer().transform?.(createMinimalMockedTransformOptions());
+
+    expect(MockDefaultBabelTransformer.transform).toHaveBeenCalledTimes(1);
+    expect(MockDefaultBabelTransformer.transform).toHaveBeenCalledWith(
+      expect.objectContaining({
+        plugins: expect.arrayContaining([expect.objectContaining({ name: 'componentNameAnnotatePlugin' })]),
+      }),
+    );
   });
 
   test.each([
@@ -70,7 +114,7 @@ describe('SentryBabelTransformer', () => {
       } as BabelTransformerArgs,
     ],
   ])('transform does not add plugin if filename includes node_modules', input => {
-    SentryBabelTransformer.transform?.(input);
+    createSentryBabelTransformer().transform?.(input);
 
     expect(MockDefaultBabelTransformer.transform).toHaveBeenCalledTimes(1);
     expect(MockDefaultBabelTransformer.transform).toHaveBeenCalledWith({
@@ -79,3 +123,13 @@ describe('SentryBabelTransformer', () => {
     });
   });
 });
+
+function createMinimalMockedTransformOptions(): BabelTransformerArgs {
+  return {
+    filename: '/project/file',
+    options: {
+      projectRoot: 'project/root',
+    },
+    plugins: [],
+  } as BabelTransformerArgs;
+}